Restructured article form folder/file structure.
