+++ app/javascript/src/views/ArticleForm/components/BodyMarkdown.jsx
