+++ app/javascript/src/views/ArticleForm/components/BodyPreview.jsx
