+++ app/javascript/src/views/ArticleForm/components/ImageManagement.jsx
