+++ app/javascript/src/views/ArticleForm/components/MainImage.jsx
