+++ app/javascript/src/views/ArticleForm/components/Notice.jsx
