+++ app/javascript/src/views/ArticleForm/components/PublishToggle.jsx
